Updated Architecture
Moved puzzle specs that are generated from the XML to its own class (PuzzleSpecs) for central access.

Moved board logic to its own class (Board)

Created a main.py file for parsing the command line and making calls to the appropriate classes and their functions for making and solving a puzzle.
